+++ PA2/main.c
@@ -57,7 +57,7 @@ int main(int argc, char *argv[])
     // Allocate the memory for the target.
     host_c.shape[0] = host_a.shape[0];
     host_c.shape[1] = host_a.shape[1];
-    host_c.data = (float *)malloc(sizeof(float) * host_c.shape[0] * host_c.shape[1]);
+    host_c.data = (int *)malloc(sizeof(int) * host_c.shape[0] * host_c.shape[1]);
 
     // Find platforms and devices
     OclPlatformProp *platforms = NULL;
